Modified def flash_button to allow timer of 0
Using def flash_button with a timer of 0 allows the game to
mark buttons as "checked" quicker that it could before, as
previously it had to wait for each flash to complete before
moving on to the next button.

New version allows def check_adjacent to run significantly
quicker and with less data processed per button.