Modified def check_adjacent for better results
Also added def reveal_borders to make user experience align
closer to the real game.  Now when clicking a button with 0
adjacent mined buttons, it will reveal all adjacent buttons 0
adjacent mined buttons, then reveal the buttons adjacent to
the just-revealed buttons.

The end result is that the game functions smoother and quicker
with less user input required, while keeping the gameplay
essentially the same as it was before.